[project @ test.consumer: added tests for consumer.consumer.GenericConsumer._checkNonce]
This replaces test.consumer.NonceIdResTest, which was not actually testing
anything related to nonces at all.  (While it's true that each attempt was
resulting in a Failure result, the failure was due to the lack of a signature,
and not anything to do with the properties of the nonces.)

I didn't preserve test_twoNonce, shout if that's a problem.