[project @ [API] server.CheckIDRequest.answer: include op_endpoint in response [#2221]] server.Server: take op_endpoint in constructor server.CheckIDRequest.fromMessage: also take Server
The story: 

There were several things about the URL argument when it first
appeared in the signature to CheckIDRequest.answer.  It was only
required in this weird little corner case, and we weren't really sure
how it was going to be used, and user_setup_url might have been
different every time.  So I added it to answer() and made it optional
except in the case where it actually needed to be there, and if people
thought it was weird, well, maybe they'd chime in and help remove it
from the protocol.

Then, based on feedback from library users, it changed from being the
user_setup_url to just the server URL, which we would use to construct
a plausable-looking user_setup_url from, as nobody was going to do
anything else with it anyway.  But it was still optional but for that
one corner case.

Now we need the URL for *every* id_res response, and it's the server's
endpoint URL, not some weird poorly-understood thing that might vary
with each request.  Given that it's the server's endpoint URL, I'm
putting it on the Server object, and doing some internal
reference-passing where needed.  (I hope I didn't create any reference
cycles, but given that Servers don't hold references to their
Requests, I'm pretty sure we're okay.)


The result:

Server() takes a second parameter, op_endpoint, which is optional for
backwards compatibility but is actually required if you ever want to
reply to a version 2.0 message.  It uses warnings.warn to throw a
warning if you leave it out.

OpenIDRequest.fromMessage takes a second parameter, server, which most
of the concrete implementations of OpenIDRequest don't use, but is not
optional for CheckIDRequest.  This does break compatibility, but I'll
be pretty surprised if anyone is using fromMessage() directly instead
of going through Decoder.decode().