[project @ Be smarter about DB API exceptions]
Python 2.5 has sqlite3 built-in, and we ran into a Python install that
thus had sqlite3 and pysqlite2, which are really the same thing. The
problem with this is that there's a good chance that we'll use the
exception classes from the wrong module and so the except clauses
won't actually trap any exceptions. The sqlite modules both use a DB
API extension that helps us, putting the exception classes as
attributes on the connection. psycopg2 does too. MySQLdb does
not. This patch sorts it out, and also tries to tell the user earlier
if something goes wrong.