Fix normalization of non-ASCII query strings on Python 2
urinorm currently deals with encoding issues when normalizing the path,
but not the query string.  However, in some cases it can happen that the
query string contains non-ASCII characters, particularly if using
https://openid.net/specs/openid-simple-registration-extension-1_0.html
in which case the user's full name may very well not be entirely ASCII;
on Python 2 this resulted in a UnicodeEncodeError in urlencode.  Work
around this.